final_project/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
joblib==1.2.0
numexpr==2.8.4
numpy==1.23.5
packaging==21.3
pandas==1.5.1
pickle-mixin==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
tables==3.7.0
threadpoolctl==3.1.0
urllib3==1.26.12
python-dotenv==0.21.0
